+++ AndroidApp/app/src/main/java/me/nya_n/notificationnotifier/App.kt
@@ -5,13 +5,13 @@ import me.nya_n.notificationnotifier.data.repository.AppRepository
 import me.nya_n.notificationnotifier.data.repository.UserSettingRepository
 import me.nya_n.notificationnotifier.data.repository.source.DB
 import me.nya_n.notificationnotifier.data.repository.source.UserSettingDataStore
-import me.nya_n.notificationnotifier.domain.*
+import me.nya_n.notificationnotifier.domain.usecase.*
+import me.nya_n.notificationnotifier.domain.util.SharedPreferenceProvider
 import me.nya_n.notificationnotifier.ui.screen.MainViewModel
 import me.nya_n.notificationnotifier.ui.screen.SharedViewModel
 import me.nya_n.notificationnotifier.ui.screen.detail.DetailViewModel
 import me.nya_n.notificationnotifier.ui.screen.selection.SelectionViewModel
 import me.nya_n.notificationnotifier.ui.screen.top.TopViewModel
-import me.nya_n.notificationnotifier.utils.SharedPreferenceProvider
 import org.koin.android.ext.koin.androidContext
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.core.context.startKoin
