+++ AndroidApp/domain/src/main/java/me/nya_n/notificationnotifier/domain/usecase/DeleteTargetAppUseCase.kt
@@ -1,15 +1,15 @@
-package me.nya_n.notificationnotifier.domain
+package me.nya_n.notificationnotifier.domain.usecase
 
 import me.nya_n.notificationnotifier.data.repository.AppRepository
 import me.nya_n.notificationnotifier.model.InstalledApp
 
 /**
- * 通知対象に追加
+ * 通知対象から外す
  */
-class AddTargetAppUseCase(
+class DeleteTargetAppUseCase(
     private val appRepository: AppRepository
 ) {
     suspend operator fun invoke(target: InstalledApp) {
-        appRepository.addTargetApp(target)
+        appRepository.deleteTargetApp(target)
     }
 }
