+++ AndroidApp/domain/src/main/java/me/nya_n/notificationnotifier/domain/usecase/ImportDataUseCase.kt
@@ -1,4 +1,4 @@
-package me.nya_n.notificationnotifier.domain
+package me.nya_n.notificationnotifier.domain.usecase
 
 import android.content.Context
 import android.net.Uri
