+++ AndroidApp/domain/src/main/java/me/nya_n/notificationnotifier/domain/usecase/LoadFilterConditionUseCase.kt
@@ -1,4 +1,4 @@
-package me.nya_n.notificationnotifier.domain
+package me.nya_n.notificationnotifier.domain.usecase
 
 import me.nya_n.notificationnotifier.data.repository.AppRepository
 import me.nya_n.notificationnotifier.model.InstalledApp
