+++ AndroidApp/ui/src/main/java/me/nya_n/notificationnotifier/ui/screen/detail/DetailViewModel.kt
@@ -3,9 +3,9 @@ package me.nya_n.notificationnotifier.ui.screen.detail
 import android.content.Context
 import androidx.lifecycle.*
 import kotlinx.coroutines.launch
-import me.nya_n.notificationnotifier.domain.DeleteTargetAppUseCase
-import me.nya_n.notificationnotifier.domain.LoadFilterConditionUseCase
-import me.nya_n.notificationnotifier.domain.SaveFilterConditionUseCase
+import me.nya_n.notificationnotifier.domain.usecase.DeleteTargetAppUseCase
+import me.nya_n.notificationnotifier.domain.usecase.LoadFilterConditionUseCase
+import me.nya_n.notificationnotifier.domain.usecase.SaveFilterConditionUseCase
 import me.nya_n.notificationnotifier.model.InstalledApp
 import me.nya_n.notificationnotifier.model.Message
 import me.nya_n.notificationnotifier.ui.R
