+++ AndroidApp/ui/src/main/java/me/nya_n/notificationnotifier/ui/screen/selection/SelectionViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.launch
-import me.nya_n.notificationnotifier.domain.AddTargetAppUseCase
+import me.nya_n.notificationnotifier.domain.usecase.AddTargetAppUseCase
 import me.nya_n.notificationnotifier.model.InstalledApp
 import me.nya_n.notificationnotifier.model.Message
 import me.nya_n.notificationnotifier.ui.R
